--- a/index.html
+++ b/index.html
@@ -11,6 +11,7 @@
       href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,400;0,700;1,400;1,700&display=swap"
       rel="stylesheet"
     />
+    <link rel="preload" as="image" type="image/webp" imagesrcset="./static/hero.webp 375w" />
     <title>memegle - gif search engine for you</title>
   </head>