--- a/index.html
+++ b/index.html
@@ -9,7 +9,9 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,400;0,700;1,400;1,700&display=swap"
-      rel="stylesheet"
+      rel="preload"
+      media="print"
+      onload="this.media='all'"
     />
     <link rel="preload" as="image" type="image/webp" imagesrcset="./static/hero.webp 375w" />
     <title>memegle - gif search engine for you</title>