--- a/package.json
+++ b/package.json
@@ -41,6 +41,7 @@
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "^6.2.0",
+    "css-minimizer-webpack-plugin": "^7.0.0",
     "dotenv-webpack": "^7.0.3",
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.5.0",
@@ -53,12 +54,13 @@
     "file-loader": "^6.2.0",
     "html-loader": "^2.1.2",
     "html-webpack-plugin": "^5.3.2",
+    "mini-css-extract-plugin": "^2.9.1",
     "prettier": "^2.3.2",
     "style-loader": "^3.2.1",
     "ts-loader": "^9.3.1",
     "typescript": "^4.8.2",
-    "webpack": "^5.88.2",
-    "webpack-cli": "^4.7.2",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^3.11.2"
   },
   "babel": {