--- a/package.json
+++ b/package.json
@@ -1,7 +1,9 @@
 {
   "name": "frontend-performance-basecamp",
   "version": "0.0.0",
-  "sideEffects": false,
+  "sideEffects": [
+    "*.css"
+  ],
   "engines": {
     "node": ">=18.0.0"
   },