--- a/src/pages/Home/Home.module.css
+++ b/src/pages/Home/Home.module.css
@@ -4,6 +4,10 @@
   width: 100%;
   height: 46.5rem;
   padding-top: 3.5rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 .heroImage {