--- a/src/pages/Home/components/CustomCursor/CustomCursor.module.css
+++ b/src/pages/Home/components/CustomCursor/CustomCursor.module.css
@@ -1,11 +1,12 @@
 .cursor {
   pointer-events: none;
-  position: absolute;
+  position: fixed;
   top: 0;
   left: 0;
   z-index: 10;
   padding: 0 16px;
   color: var(--white);
+  will-change: transform;
 }
 
 .cursor::before {