--- a/src/pages/Home/components/CustomCursor/CustomCursor.module.css
+++ b/src/pages/Home/components/CustomCursor/CustomCursor.module.css
@@ -1,6 +1,6 @@
 .cursor {
   pointer-events: none;
-  position: fixed;
+  position: absolute;
   top: 0;
   left: 0;
   z-index: 10;