--- a/src/pages/Home/components/FeatureItem/FeatureItem.tsx
+++ b/src/pages/Home/components/FeatureItem/FeatureItem.tsx
@@ -8,7 +8,7 @@ type FeatureItemProps = {
 const FeatureItem = ({ title, imageSrc }: FeatureItemProps) => {
   return (
     <div className={styles.featureItem}>
-      <img className={styles.featureImage} src={imageSrc} />
+      <video className={styles.featureImage} src={imageSrc} loop muted playsInline autoPlay />
       <div className={styles.featureTitleBg}></div>
       <h4 className={styles.featureTitle}>{title}</h4>
     </div>