--- a/src/pages/Search/components/GifItem/GifItem.module.css
+++ b/src/pages/Search/components/GifItem/GifItem.module.css
@@ -9,12 +9,12 @@
     rgba(232, 109, 198, 0.1) 97.71%
   );
   overflow: hidden;
-  transition: all 0.2s ease-in;
   border-radius: 4%;
+  transition: transform 0.2s ease;
 }
 
 .gifItem:hover {
-  top: -0.75rem;
+  transform: translateY(-0.75rem);
 }
 
 .gifImage {