--- a/src/pages/Search/components/GifItem/GifItem.tsx
+++ b/src/pages/Search/components/GifItem/GifItem.tsx
@@ -17,3 +17,5 @@ const GifItem = ({ imageUrl = '', title = '' }: GifItemProps) => {
 };
 
 export default GifItem;
+
+GifItem.displayName = 'GifItem';