--- a/src/pages/Search/components/GifItem/GifItem.tsx
+++ b/src/pages/Search/components/GifItem/GifItem.tsx
@@ -1,3 +1,4 @@
+import { memo } from 'react';
 import { GifImageModel } from '../../../../models/image/gifImage';
 
 import styles from './GifItem.module.css';
@@ -16,6 +17,6 @@ const GifItem = ({ imageUrl = '', title = '' }: GifItemProps) => {
   );
 };
 
-export default GifItem;
+export default memo(GifItem);
 
 GifItem.displayName = 'GifItem';