--- a/src/pages/Search/components/HelpPanel/HelpPanel.module.css
+++ b/src/pages/Search/components/HelpPanel/HelpPanel.module.css
@@ -10,7 +10,7 @@
 
 .selectedItemContainer {
   position: fixed;
-  right: -320px;
+  right: 0;
   top: 0;
   width: 320px;
   height: 100%;
@@ -20,11 +20,13 @@
   backdrop-filter: blur(5px);
   -webkit-backdrop-filter: blur(5px);
   opacity: 0;
-  transition: 0.5s all cubic-bezier(0.82, 0.085, 0.395, 0.895);
+  transform: translateX(100%);
+  transition: transform 0.5s cubic-bezier(0.82, 0.085, 0.395, 0.895),
+    opacity 0.5s cubic-bezier(0.82, 0.085, 0.395, 0.895);
 }
 
 .selectedItemContainer.showSheet {
-  right: 0;
+  transform: translateX(0);
   opacity: 1;
 }
 
@@ -41,11 +43,12 @@
 
 .sheetContentsContainer {
   padding: 20px 16px 80px 16px;
-  height: 100%;
+  height: calc(100% - 3.5rem);
   line-height: 1.5rem;
-  overflow-y: scroll;
+  overflow-y: auto;
 }
 
 .sheetContentsContainer img {
   width: 100%;
+  height: auto;
 }