--- a/src/types/images.d.ts
+++ b/src/types/images.d.ts
@@ -2,3 +2,5 @@ declare module '*.png';
 declare module '*.jpg';
 declare module '*.gif';
 declare module '*.svg';
+declare module '*.webp';
+declare module '*.webm';