--- a/webpack.config.js
+++ b/webpack.config.js
@@ -58,6 +58,8 @@ module.exports = {
   },
   optimization: {
     minimize: true,
-    minimizer: [new CssMinimizerPlugin(), '...']
+    minimizer: [new CssMinimizerPlugin(), '...'],
+    usedExports: true,
+    sideEffects: true
   }
 };