… src/assets/images/find.gif
… src/assets/images/find.webm
… src/assets/images/free.gif
… src/assets/images/free.webm
… src/assets/images/hero.webp
… src/assets/images/trending.gif
… src/assets/images/trending.webm