버건디 안녕하세요! 초코입니다.
처음으로 성능 개선을 해보느라 새로운 학습 키워드들이 많았던 것 같은데요, 코드 리뷰 과정에서 부족한 점을 채워나갈 수 있으면 좋겠어요. 코드 리뷰 잘 부탁드려요 😄 

<br/>

## 🔥 결과

<!-- 개선 목표에 있는 측정 항목들에 대해 개선 작업 전/후의 성능 측정 결과를 적어주세요. -->

- 배포한 CloudFront 접근 경로: https://d1oy9kvngke82j.cloudfront.net

<br/>

- 개선 전후 성능 측정 결과
  - 성능 개선 기준
    - Lighthouse performance `95점` 이상
    - Home 페이지에서 불러오는 스크립트 리소스 크기 `< 60kb`
    - 히어로 이미지 크기 `< 120kb`
    - `프랑스 파리`에서 `Fast 3G` 환경으로 접속했을 때 => 파리는 대기열이 너무 길어서 '네덜란드' 기준으로 진행했습니다.
        - Home 첫 번째 로드시 LCP `< 2.5s`
        - Home 두 번째 이후 로드시 LCP `< 1.5s`

<br/>

  - 개선 전(GitHub Pages) : [사이트 바로가기](https://00kang.github.io/perf-basecamp/)
  
    <table>
      <tr>
        <th width="30%">항목</th>
        <th width="10%">결과</th>
        <th width="60%">참고 이미지</th>
      </tr>
      <tr>
        <td>Lighthouse performance</td>
        <td>70점</td>
        <td><img src="https://github.com/user-attachments/assets/343bd94d-5177-4624-9a14-78048ab964a5" /></td>
      </tr>
      <tr>
        <td>Home 페이지에서 불러오는 스크립트 리소스 크기</td>
        <td>253 kB</td>
        <td><img src="https://github.com/user-attachments/assets/6acca5cf-6d9e-4ded-a843-a60fc8176bbd"/></td>
      </tr>
      <tr>
        <td>히어로 이미지 크기</td>
        <td>10.7 MB</td>
        <td></td>
      </tr>
      <tr>
        <td>네덜란드에서 Fast 3G 환경으로 접속했을 때, Home 첫 번째 로드 시 LCP</td>
        <td>2.511s</td>
        <td><img src="https://github.com/user-attachments/assets/f5c840e1-26dc-4c5c-9636-75e80f51c010" /></td>
      </tr>
      <tr>
        <td>네덜란드에서 Fast 3G 환경으로 접속했을 때, Home 두 번째 이후 로드시 LCP</td>
        <td>0.330s</td>
        <td></td>
      </tr>
    </table>

<br/>

  - 개선 후(CloudFront) : [사이트 바로가기](https://d1oy9kvngke82j.cloudfront.net)
 
    <table>
      <tr>
        <th width="30%">항목</th>
        <th width="10%">결과</th>
        <th width="60%">참고 이미지</th>
      </tr>
      <tr>
        <td>Lighthouse performance</td>
        <td>99점</td>
        <td><img src="https://github.com/user-attachments/assets/1085c3e7-7ef7-45a6-8e2f-dac72b748a24" /></td>
      </tr>
      <tr>
        <td>Home 페이지에서 불러오는 스크립트 리소스 크기</td>
        <td>49.0 kB</td>
        <td><img src="https://github.com/user-attachments/assets/31717ef6-814d-461f-aaac-97e3354325f4"/></td>
      </tr>
      <tr>
        <td>히어로 이미지 크기</td>
        <td>101 kB</td>
        <td></td>
      </tr>
      <tr>
        <td>네덜란드에서 Fast 3G 환경으로 접속했을 때, Home 첫 번째 로드 시 LCP</td>
        <td>2.335s</td>
        <td><img src="https://github.com/user-attachments/assets/6953891b-86dd-4665-a927-3b2af99d34f5"/></td>
      </tr>
      <tr>
        <td>네덜란드에서 Fast 3G 환경으로 접속했을 때, Home 두 번째 이후 로드시 LCP</td>
        <td>1.353s</td>
        <td></td>
      </tr>
    </table>

<br/>

## ✅ 개선 작업 목록

<!-- 각 요구사항을 위해 어떤 개선 작업을 진행했는지 적어주세요
코드 변경사항으로 확인하기 어려운 CloudFront 설정 사항 등은 리뷰어가 확인할 수 있게 스크린샷이나 적용한 항목들을 적어주면 좋겠지요? 🙂
-->


**1 요청 크기 줄이기**

- [x] 소스코드 크기 줄이기
  - Webpack v5 부터는 빌드할 때 내장된 TerserPlugin을 통해 자동으로 js 코드를 minify하고 uglify 합니다. 따라서 **MiniCssExtractPlugin**, **CssMinimizerPlugin**로 css에 한해서만 진행했습니다.
  - bundle 사이즈를 줄이기 위해 압축 파일인 **gzip**를 사용했습니다.
  
- [x] 이미지 크기 줄이기
  - 사이트에서 사용되는 이미지의 개수가 적기 때문에 별도의 플러그인을 이용하기 보다는 외부 사이트를 통해 gif 이미지는 **webm** 이미지로, png 이미지는 **webp** 이미지로 변경하여 크기를 줄였습니다.
  
  <img src="https://github.com/user-attachments/assets/86e810a6-f3a2-47d2-8652-ddc123a9c221" width="300px" />

<br/>

**2 필요한 것만 요청하기**

- [x] 페이지별 리소스 분리
  - Home과 Search 페이지를 분리했습니다.
  
- [x] 아이콘 패키지 Tree Shaking
  - react-icons : react-icons로 불필요한 아이콘까지 불러오던 것에서 **react-icons/all-files**로 변경하여 사용하는 아이콘만 정확하게 불러오도록 변경했습니다.
  
  |react-icons|@react-icons/all-files|
  |------------|------------|
  |<img src="https://github.com/user-attachments/assets/67d773ad-29b9-43ab-9836-b739e9c17825" width="300px" /> |<img src="https://github.com/user-attachments/assets/cd37b760-f7da-43a3-be18-cf020fc69e47" width="300px" />  |

<br/>

**3 같은 건 매번 새로 요청하지 않기**

- [x] CloudFront 캐시 설정 / S3 메타데이터 설정 (설정값, 해당 값을 설정한 이유 포함)

  |응답헤더|정책|
  |---|---|
  |<img src="https://github.com/user-attachments/assets/d10d2eba-d9db-49c0-9ed3-c1fb1a7ddd50" /> | <img src="https://github.com/user-attachments/assets/80996bed-81ad-48b1-ba48-7101ee981a21" /> 

- [x] GIPHY의 trending API를 Search 페이지에 들어올 때마다 새로 요청하지 않아야 한다.
  - fetchGifs 함수를 **fetchGifsWithCache** 함수로 변경하여 캐싱 기능을 추가했습니다. 메모리 내 캐시를 구현하여 동일한 요청에 대해 일정 시간 동안 결과를 재사용합니다.

<br/>

**4 최소한의 변경만 일으키기**

- [x] 검색 결과 > 추가 로드시 추가된 목록만 새로 렌더되어야 한다.
  - 불필요한 리렌더링을 방지하기 위해 **React.memo**를 추가했습니다. imageUrl과 title prop이 변경되지 않는 한, GifItem 컴포넌트는 리렌더링되지 않습니다.

  |before|after|
  |---|---|
  |<img src="https://github.com/user-attachments/assets/04a2f498-7a36-483f-80e9-0f566fbe9edf" /> | <img src="https://github.com/user-attachments/assets/0050ee54-6469-4716-bd95-b9743a9208de" /> 

- [x] Layout Shift 없이 애니메이션이 일어나야 한다.

  - **custom cursor** 
    - transform 속성 활용
  - **검색결과>hover**
    - top 속성 대신 transform: translateY(-0.75rem) 사용
    - transition 속성을 transform에만 적용
  - **도움말 패널 열고닫기 애니메이션**
    - right 속성 대신 transform: translateX(100%)를 사용
    - transition 속성을 transform과 opacity에만 적용
  

- [x] Frame Drop이 일어나지 않아야 한다.
  - (Chrome DevTools 기준) Partially Presented Frame 역시 최소로 발생해야 한다.

<br/>

## 🧐 공유

<!-- 작업하면서 든 생각, 질문, 새롭게 학습하거나 시도해본 내용 등등 공유할 사항이 있다면 자유롭게 적어주세요 -->
- webpagetest는 '파리'가 아니라 '네덜란드'를 기준으로 진행했습니다.
- 네덜란드에서 Fast 3G 환경으로 접속했을 때, Home 첫 번째 로드 시 LCP는 줄었으나, 오히려 두 번째 이후 로드시 LCP는 늘어나는 현상이 발견되었습니다. 왜 그럴까용? ❓